requirements.txt
addict==2.4.0
efficientnet-pytorch==0.7.1
importlib-metadata==7.0.1
joblib==1.3.2
mmcv-full==1.2.0
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.2.0
platformdirs==4.1.0
PyWavelets==1.4.1
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
yapf==0.40.2
zipp==3.17.0
